+++ tests/commands/account-balance-2.sh
@@ -1,1 +1,0 @@
-"$QUILL" account-balance --pem-file - --dry-run