+++ tests/commands/ckbtc-retrieve-btc.sh
@@ -1,1 +1,0 @@
-"$QUILL" ckbtc retrieve-btc 3L2Uyh1eHpfPyPayqrh5WjfnTzWiG4xPLu --amount 3.14 --memo 9 --pem-file - | "$QUILL" send --dry-run -