+++ tests/commands/neuron-manage-a.sh
@@ -1,1 +1,0 @@
-"$QUILL" neuron-manage 2313380519530470538 -a 3600 --pem-file - | "$QUILL" send --dry-run -