+++ tests/commands/neuron-manage-add-hot-key.sh
@@ -1,1 +1,0 @@
-"$QUILL" neuron-manage 2313380519530470538 --pem-file - --add-hot-key fdsgv-62ihb-nbiqv-xgic5-iefsv-3cscz-tmbzv-63qd5-vh43v-dqfrt-pae | "$QUILL" send --dry-run -