+++ tests/commands/neuron-manage-clear.sh
@@ -1,1 +1,0 @@
-"$QUILL" neuron-manage 2313380519530470538 --clear-manage-neuron-followees --pem-file - | "$QUILL" send --dry-run -