+++ tests/commands/neuron-manage-disburse-stop-dissolving.sh
@@ -1,1 +1,0 @@
-"$QUILL" neuron-manage 2313380519530470538 --disburse --stop-dissolving --pem-file - | "$QUILL" send --dry-run -