+++ tests/commands/neuron-manage-merge.sh
@@ -1,1 +1,0 @@
-"$QUILL" neuron-manage 2313380519530470538 --merge-from-neuron 380519530470538 --pem-file - | "$QUILL" send --dry-run -