+++ tests/commands/neuron-manage-vote.sh
@@ -1,1 +1,0 @@
-"$QUILL" neuron-manage 2313380519530470538 --register-vote 123 456 --pem-file - | "$QUILL" send --dry-run -