+++ tests/commands/public-ids.sh
@@ -1,2 +1,0 @@
-"$QUILL" public-ids --pem-file -
-"$QUILL" public-ids --principal-id 44mwt-bq3um-tqicz-bwhad-iipx4-6wzex-olvaj-z63bj-wkelv-xoua3-rqe