+++ tests/commands/sns-list-deployed-snses.sh
@@ -1,1 +1,0 @@
-"$QUILL" sns list-deployed-snses --canister-ids-file ./sns_canister_ids.json --dry-run