+++ tests/commands/sns-neuron-stake-no-amount.sh
@@ -1,1 +1,0 @@
-"$QUILL" sns stake-neuron --memo 777 --claim-only --canister-ids-file ./sns_canister_ids.json --pem-file - | "$QUILL" send --dry-run -