+++ tests/commands/sns-register-vote-no.sh
@@ -1,3 +1,0 @@
-NEURON_ID=83a7d2b12f654ff58335e5a2512ccae0d7839c744b1807a47c96f5b9f3969069
-"$QUILL" sns register-vote $NEURON_ID --proposal-id 1 --vote n --canister-ids-file ./sns_canister_ids.json --pem-file - | "$QUILL" send --dry-run -
-